% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_change_dir.R
\name{sbn_change_dir}
\alias{sbn_change_dir}
\title{Convert direction of SBN}
\usage{
sbn_change_dir(g, method = c("rev", "undir"))
}
\arguments{
\item{g}{a river network as an igraph object}

\item{method}{One of "rev" or "undir", determining what to convert the network to.}
}
\value{
An igraph SBN
}
\description{
Convert a downstream directed SBN to either, reversed or undirected.
}
\examples{
\dontrun{
sbn_direction(g, method = "up")
}
}
