% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_create.R
\name{sbn_create}
\alias{sbn_create}
\title{Create SBNs}
\usage{
sbn_create(n, p)
}
\arguments{
\item{n}{integer number of nodes.}

\item{p}{branching probability, from 0 - 1}
}
\value{
An adjacency matrix
}
\description{
Creates an SBN representation of a river
}
\examples{
sbn_create(10, 0.5)

}
