% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_downstream_vert.R, R/sbn_upstream_vert.R
\name{sbn_downstream_vert}
\alias{sbn_downstream_vert}
\title{find all vertices downstream of a given vertex}
\usage{
sbn_downstream_vert(g, vert)

sbn_downstream_vert(g, vert)
}
\arguments{
\item{g}{an SBN as an igraph object. Must be a downstream directed graph.}

\item{vert}{target vertex to get all upstream nodes of}
}
\value{
a vector of downstream vertex id's

a vector of upstream vertex id's
}
\description{
Find all vertices downstream of a given vertex.

Find all vertices upstream of a given vertex.
}
\examples{
\dontrun{
sbn_downstream_vert(g, 5)
}
\dontrun{
sbn_upstream_vert(g, 5)
}
}
