% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_dwn_nwk.R
\name{sbn_dwn_nwk}
\alias{sbn_dwn_nwk}
\title{Convert to a downstream directed network}
\usage{
sbn_dwn_nwk(g, mouth)
}
\arguments{
\item{g}{a river network as an igraph object}

\item{mouth}{the rivermouth vertex}
}
\value{
a downstream directed network.
}
\description{
Convert a non-directed network to a downstream directed network.
}
\examples{
\dontrun{
g <- sbn_create(10, 0.7)
g2 <- igraph::as.undirected(g)
sbn_dwn_nwk(g2, mouth = 1)
}
}
