% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_downstream.R
\name{sbn_get_downstream}
\alias{sbn_get_downstream}
\title{Find all downstream nodes}
\usage{
sbn_get_downstream(g, node)
}
\arguments{
\item{g}{A river network as an igraph object. Must be a downstream directed graph.}

\item{node}{target node to get all downstream nodes of.}
}
\value{
a vector of downstream node id's.
}
\description{
Find all nodes downstream of a given node.
}
\examples{
g <- sbn_create(10, 0.5)
sbn_get_downstream(g, 10)

}
