% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_hw.R
\name{sbn_get_hw}
\alias{sbn_get_hw}
\title{Find all headwater nodes}
\usage{
sbn_get_hw(g)
}
\arguments{
\item{g}{A river network as an igraph object. Must be a downstream directed graph.}
}
\value{
A vector of headwater node id's.
}
\description{
Find all headwater nodes in a network.
}
\examples{
g <- sbn_create(10, 0.5)
sbn_get_hw(g)

}
