% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_outlet.R
\name{sbn_get_outlet}
\alias{sbn_get_outlet}
\title{Find river mouth vertex}
\usage{
sbn_get_outlet(g)
}
\arguments{
\item{g}{a igraph directed graph representation of a river, e.g. an SBN or OCN.}
}
\value{
an integer identifying the river mouth  vertex
}
\description{
Find river mouth node from a directed graph.
}
\examples{
\dontrun{
sbn_get_outlet(g)
}
}
