% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_upstream.R
\name{sbn_get_upstream}
\alias{sbn_get_upstream}
\title{find all vertices downstream of a given vertex}
\usage{
sbn_get_upstream(g, vert)
}
\arguments{
\item{g}{an SBN as an igraph object. Must be a downstream directed graph.}

\item{vert}{target vertex to get all upstream nodes of}
}
\value{
a vector of upstream vertex id's
}
\description{
Find all vertices upstream of a given vertex.
}
\examples{
\dontrun{
sbn_get_upstream(g, 5)
}
}
