% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_hw_nodes.R
\name{sbn_hw_nodes}
\alias{sbn_hw_nodes}
\title{get all headwater nodes}
\usage{
sbn_hw_nodes(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
a vector of headwater node id's
}
\description{
get all headwater nodes
}
\examples{
\dontrun{
sbn_hw_nodes(g)
}
}
