% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_to_adj.R
\name{sbn_to_mtx}
\alias{sbn_to_mtx}
\title{Convert SBN to various adjacency matrix formats}
\usage{
sbn_to_mtx(
  g,
  method = c("dwn_mtx", "undir_mtx", "up_mtx", "v2v_dist_up", "v2v_dist_dwn",
    "v2v_dist_undir", "v2v_dist_dwn"),
  unconnected = Inf,
  weights = NULL
)
}
\arguments{
\item{g}{An SBN}

\item{method}{One of "dwn_mtx", an adjacency matrix for a downstream directed SBN, "up_mtx", an adjacency matrix for a upstream directed SBN, "undir_mtx", an adjacency matrix for a  undirected SBN or "v2v_dist", an adjacency matrix of vertex to vertex distances.}

\item{unconnected}{When generating vertext-to-vertex distance matrices, what value should be used for unconnected elements. For example, in a downstream directed network, all upstream links are considered unconnected. Default value is \code{Inf} but other options are possible, such as \code{NA} or \code{0}.}

\item{weights}{Passed to \code{igraph::shortest.paths()}. Possibly a numeric vector giving edge weights. If this is \verb{NULL`` and the graph has a weight edge attribute, then the attribute is used. If this is }NA` then no weights are used (even if the graph has a weight attribute).}
}
\value{
An adjacency or distance matrix
}
\description{
Convert a downstream directed SBN to various adjacency matrix or distance matrix formats.
}
\examples{
\dontrun{
sbn_to_mtx(g, method = "dwn_mtx")
}
}
