% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_change_dir.R
\name{sbn_change_dir}
\alias{sbn_change_dir}
\title{Change the upstream/downstream direction of an SBN}
\usage{
sbn_change_dir(g, method = c("rev", "undir"))
}
\arguments{
\item{g}{A river network as an igraph object. Must be a downstream directed graph.}

\item{method}{One of "rev" or "undir", determining what to convert the network to.}
}
\value{
A river network as an igraph object.
}
\description{
Change the upstream/downstream direction of an SBN to either, reversed or undirected.
}
\examples{
g <- sbn_create(10, 0.7)
sbn_change_dir(g, method = "rev")

}
