% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_down_dir.R
\name{sbn_down_dir}
\alias{sbn_down_dir}
\title{Convert to a downstream directed network}
\usage{
sbn_down_dir(g, mouth)
}
\arguments{
\item{g}{a river network as an igraph object}

\item{mouth}{the rivermouth vertex}
}
\value{
a downstream directed network.
}
\description{
Convert a non-directed network to a downstream directed network.
}
\examples{
\dontrun{
g <- sbn_create(10, 0.7)
g2 <- igraph::as.undirected(g)
sbn_down_dir(g2, mouth = 1)
}
}
