% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_downstream.R
\name{sbn_get_downstream}
\alias{sbn_get_downstream}
\title{find all vertices downstream of a given vertex}
\usage{
sbn_get_downstream(g, vert)
}
\arguments{
\item{g}{an SBN as an igraph object. Must be a downstream directed graph.}

\item{vert}{target vertex to get all downstream nodes of}
}
\value{
a vector of downstream vertex id's
}
\description{
find all vertices downstream of a given vertex
}
\examples{
\dontrun{
sbn_get_downstream(g, 5)
}
}
