% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_hw.R
\name{sbn_get_hw}
\alias{sbn_get_hw}
\title{get all headwater nodes}
\usage{
sbn_get_hw(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
a vector of headwater node id's
}
\description{
get all headwater nodes
}
\examples{
\dontrun{
sbn_get_hw(g)
}
}
