% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_outlet.R
\name{sbn_get_outlet}
\alias{sbn_get_outlet}
\title{Find river mouth node}
\usage{
sbn_get_outlet(g)
}
\arguments{
\item{g}{A river network as an igraph object. Must be a downstream directed graph.}
}
\value{
An integer identifying the id of river mouth node.
}
\description{
Find river mouth node from a directed graph.
}
\examples{
g <- sbn_create(10, 0.5)
sbn_get_outlet(g)

}
