% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_get_upstream.R
\name{sbn_get_upstream}
\alias{sbn_get_upstream}
\title{Find all nodes upstream of a given node}
\usage{
sbn_get_upstream(g, node)
}
\arguments{
\item{g}{a river network as an igraph object. Must be a downstream directed graph.}

\item{node}{target node to get all upstream nodes of.}
}
\value{
A vector of upstream node id's.
}
\description{
Find all nodes upstream of a given node.
}
\examples{
g <- sbn_create(10, 0.7)
sbn_get_upstream(g, 2)

}
