% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbn_strahler.R
\name{sbn_strahler}
\alias{sbn_strahler}
\title{Get strahler order}
\usage{
sbn_strahler(g)
}
\arguments{
\item{g}{a dendritic network as an igraph object}
}
\value{
a vector of stream strahler orders
}
\description{
Get strahler order
}
\examples{
\dontrun{
sbn_strahler(g)
}
}
